requirements.txt
setuptools>=65.0.0
wheel>=0.38.0
streamlit>=1.28.0
google-cloud-aiplatform>=1.73.0
PyMuPDF
pydantic
vertexai>=0.1.0
python-docx
selenium
webdriver-manager
sib-api-v3-sdk
requests
beautifulsoup4
schedule
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
reportlab
xhtml2pdf